+++ godfw-ezcsp-20160621/fib-8.go
@@ -1,7 +1,7 @@
 package main
 import "fmt"
 
-func fib(i uint64) uint64 {
+func fib(i int) int {						// HL
 
 	if i <= 2 {
 		return 1
@@ -10,7 +10,7 @@ func fib(i uint64) uint64 {
 	return fib(i - 1) + fib(i - 2)
 }
 
-func main() {
+func main() {						// HL
 
 	fmt.Println(
